storage: Fix recovery algorithms and add test for repair
